requirements.txt
fastapi
uvicorn
sqlalchemy
Rust
passlib
bcrypt
python-jose
python-multipart